+++ packages/plugin-hashtag/src/index.js
@@ -1,3 +1,2 @@
 export { createHashtagPlugin } from './createHashtagPlugin';
 export { HASHTAG_TYPE } from './types';
-export { Strategy, Component } from './decorator';