+++ packages/plugin-link/src/index.js
@@ -1,4 +1,3 @@
 export { createLinkPlugin } from './createLinkPlugin';
 export { EXTERNAL_LINK_TYPE, LINK_TYPE } from './types';
 export { default as LinkViewer } from './LinkViewer';
-export { LinkParseStrategy } from './LinkParseStrategy';