+++ packages/plugin-map/src/toolbar/MapSettingsModal.jsx
@@ -129,7 +129,7 @@ export class MapSettingsModal extends Component {
     const navBarClassName = `map_settings_modal_mobile_navbar_dynamic_${uniqueClassesId}`;
     const dividerClassName = `map_settings_modal_divider_dynamic_${uniqueClassesId}`;
 
-    /* eslint-disable camelcase */
+    /* eslint-disable camelcase*/
     this.styles.button_primary = `${buttonPrimaryClassName} ${this.styles.button_primary}`;
     this.styles.button_secondary = `${buttonSecondaryClassName} ${this.styles.button_secondary}`;
     this.styles.textInput_input = `${textInputClassName} ${this.styles.textInput_input}`;